🤺🚶‍♂️ Added unit movement

+++ TCG/Assets/_Scripts/Gameplay/Damage/Damage.cs
@@ -1,16 +1,16 @@
 public class Damage 
 {
-    float damage;
+    int damage;
     DamageSource source;
     Player playerSource;
 
-    public Damage (float damage, DamageSource source, Player playerSource) {
+    public Damage (int damage, DamageSource source, Player playerSource) {
         this.damage = damage;
         this.source = source;
         this.playerSource = playerSource;
     }
 
-    public float DamageAmount   {get {return damage;}}
+    public int DamageAmount   {get {return damage;}}
     public DamageSource Source  {get {return source;}}
     public Player PlayerSource  {get {return playerSource;}}
 }
